backend/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
python-multipart
aiofiles

# LLM & Embeddings - LangChain
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langchain-google-genai

# Vector Store & Embeddings
faiss-cpu
tiktoken

# LLM Providers
ollama
openai

# Data Processing
pymupdf
tqdm
pydantic
python-dotenv
requests

# Document Processing - for slides, notes, and office files
python-pptx
python-docx
pytesseract
Pillow

# Calendar Integration
icalendar

# Optional but recommended
certifi
